+++ CHANGELOG.md
@@ -1,11 +1,10 @@
 ## [0.1.0] - 19/11/2025
 ### Added
-- Added modify for workshops.
-
-### Added
+ - Added modify for workshops.
  - US for modify supply batch
  - US for display of supply batches based on date and idSupply
  - US for modify User with MVVM and clean architecture principles.
+ - US for qr generation
 
 ### Fixed
 - Product add necesary image